sum by mode

+++ src/main/scala/Model.scala
@@ -5,11 +5,11 @@ object Model {
 
   trait Cell
   case class UnresolvedCell(expression: Expression) extends Cell
-  case class ResolvedCell(result: Int) extends Cell
+  case class ResolvedCell(result: Long) extends Cell
 
   trait Expression
   case class CellIndex(row: Int, column: Int) extends Expression
-  case class Number(n: Int) extends Expression
+  case class Number(n: Long) extends Expression
   case class Func(name: String, range: CellRange) extends Expression
   case class BinaryOp(left: Expression, right: Expression) extends Expression
   case class CellRange(start: CellIndex, end: CellIndex)
